+++ httpclient-jetty/src/test/java/io/fabric8/kubernetes/client/jetty/JettyHttpClientProxyHttpsTest.java
@@ -13,21 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.mockwebserver.dsl;
+package io.fabric8.kubernetes.client.jetty;
 
-public interface HttpMethodable<T> {
+import io.fabric8.kubernetes.client.http.AbstractHttpClientProxyHttpsTest;
+import io.fabric8.kubernetes.client.http.HttpClient;
+import okhttp3.mockwebserver.SocketPolicy;
 
-  T any();
-
-  T post();
-
-  T get();
-
-  T put();
-
-  T delete();
-
-  T patch();
-
-  T connect();
+@SuppressWarnings("java:S2187")
+public class JettyHttpClientProxyHttpsTest extends AbstractHttpClientProxyHttpsTest {
+  @Override
+  protected HttpClient.Factory getHttpClientFactory() {
+    server.setDefaultSocketPolicy(SocketPolicy.KEEP_OPEN); // we need a challenge before switching to ssl
+    return new JettyHttpClientFactory();
+  }
 }