+++ junit/mockwebserver/src/main/java/io/fabric8/mockwebserver/utils/ResponseProviders.java
@@ -123,10 +123,12 @@ public class ResponseProviders {
 
     @Override
     public boolean equals(Object o) {
-      if (this == o)
+      if (this == o) {
         return true;
-      if (o == null || getClass() != o.getClass())
+      }
+      if (o == null || getClass() != o.getClass()) {
         return false;
+      }
 
       FixedResponseProvider<?> that = (FixedResponseProvider<?>) o;