+++ kubernetes-client-api/src/test/java/io/fabric8/kubernetes/client/http/AbstractHttpClientProxyHttpsTest.java
@@ -15,10 +15,12 @@
  */
 package io.fabric8.kubernetes.client.http;
 
+import io.fabric8.kubernetes.client.internal.SSLUtils;
 import io.fabric8.mockwebserver.DefaultMockServer;
 import io.fabric8.mockwebserver.utils.ResponseProvider;
 import okhttp3.Headers;
 import okhttp3.mockwebserver.RecordedRequest;
+import okhttp3.mockwebserver.SocketPolicy;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
@@ -26,16 +28,18 @@ import org.junit.jupiter.api.Test;
 
 import java.net.InetSocketAddress;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.Assert.assertTrue;
 
-public abstract class AbstractHttpClientProxyTest {
+public abstract class AbstractHttpClientProxyHttpsTest {
 
-  private static DefaultMockServer server;
+  protected static DefaultMockServer server;
 
   @BeforeAll
   static void beforeAll() {
-    server = new DefaultMockServer(false);
+    server = new DefaultMockServer(true);
+    server.setHttpsTunnelProxy(true);
     server.start();
   }
 
@@ -48,8 +52,9 @@ public abstract class AbstractHttpClientProxyTest {
 
   @Test
   @DisplayName("Proxied HttpClient adds required headers to the request")
-  protected void proxyConfigurationAddsRequiredHeaders() throws Exception {
-    server.expect().get().withPath("/").andReply(new ResponseProvider<String>() {
+  protected void proxyConfigurationAddsRequiredHeadersForHttps() throws Exception {
+    AtomicBoolean authenticated = new AtomicBoolean();
+    server.expect().connect().withPath("/").andReply(new ResponseProvider<String>() {
 
       @Override
       public String getBody(RecordedRequest request) {
@@ -63,7 +68,12 @@ public abstract class AbstractHttpClientProxyTest {
 
       @Override
       public int getStatusCode(RecordedRequest request) {
-        return request.getHeader(StandardHttpHeaders.PROXY_AUTHORIZATION) != null ? 200 : 407;
+        server.setDefaultSocketPolicy(SocketPolicy.UPGRADE_TO_SSL_AT_END); // for jetty to upgrade after the challenge
+        boolean auth = request.getHeader(StandardHttpHeaders.PROXY_AUTHORIZATION) != null;
+        if (auth) {
+          authenticated.set(true);
+        }
+        return auth ? 200 : 407;
       }
 
       @Override
@@ -74,18 +84,17 @@ public abstract class AbstractHttpClientProxyTest {
     }).always();
     // Given
     final HttpClient.Builder builder = getHttpClientFactory().newBuilder()
+        .sslContext(null, SSLUtils.trustManagers(null, null, true, null, null))
         .proxyAddress(new InetSocketAddress("localhost", server.getPort()))
         .proxyBasicCredentials("auth", "cred");
     try (HttpClient client = builder.build()) {
       // When
       client.sendAsync(client.newHttpRequestBuilder()
-          .uri(String.format("http://0.0.0.0:%s/not-found", server.getPort())).build(), String.class)
-          .get(10L, TimeUnit.SECONDS);
-      // Then
-      assertThat(server.getLastRequest())
-          .extracting(RecordedRequest::getHeaders)
-          .returns("0.0.0.0:" + server.getPort(), h -> h.get("Host"))
-          .returns("Basic YXV0aDpjcmVk", h -> h.get("Proxy-Authorization"));
+          .uri(String.format("https://0.0.0.0:%s/not-found", server.getPort() + 1)).build(), String.class)
+          .get(30, TimeUnit.SECONDS);
+
+      // if it fails, then authorization was not set
+      assertTrue(authenticated.get());
     }
   }
 }